API issues

+++ src/main/java/com/iemr/common/identity/data/rmnch/RMNCHBeneficiaryDetailsRmnch.java
@@ -49,11 +49,11 @@ public class RMNCHBeneficiaryDetailsRmnch {
 	@GeneratedValue
 	@Expose
 	@Column(name = "beneficiaryDetails_RmnchId", insertable = false, updatable = false)
-	private Long beneficiaryDetails_RmnchId;
+	private BigInteger beneficiaryDetails_RmnchId;
 
 	@Expose
 	@Column(name = "BeneficiaryRegID")
-	private Long BenRegId;
+	private BigInteger BenRegId;
 
 	@Expose
 	@Column(name = "aadhaNo")
@@ -289,7 +289,7 @@ public class RMNCHBeneficiaryDetailsRmnch {
 
 	@Expose
 	@Column(name = "beneficiaryId")
-	private Long benficieryid;
+	private BigInteger benficieryid;
 
 	// new fields 30-06-2021
 	@Expose
