+++ src/main/java/com/iemr/common/identity/data/rmnch/RMNCHBornBirthDetails.java
@@ -22,6 +22,7 @@
 package com.iemr.common.identity.data.rmnch;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.sql.Timestamp;
 
 import jakarta.persistence.Column;
@@ -53,11 +54,11 @@ public class RMNCHBornBirthDetails {
 
 	@Expose
 	@Column(name = "BeneficiaryRegID")
-	private Long BenRegId;
+	private BigInteger BenRegId;
 
 	@Expose
 	@Column(name = "beneficiaryId")
-	private Long benficieryid;
+	private BigInteger benficieryid;
 
 	@Expose
 	@Column(name = "countryid")
